moving necessary mappers from the `data` module to the `features:clients:data` module
